+++ README.md
@@ -9,10 +9,17 @@
 `/controllers` — папка с файлами контроллеров пользователя и карточки   
 `/models` — папка с файлами описания схем пользователя и карточки
 `/middlewares` — папка с файлами описания посредников для обработки роутов перед работой кконтроллеров
- `/models` — папка с классами различных ошибок для центролизованной системы обработки ошибок
+`/models` — папка с классами различных ошибок для центролизованной системы обработки ошибок
   
 
 ## Запуск проекта
 
 `npm run start` — запускает сервер   
 `npm run dev` — запускает сервер с hot-reload
+
+
+## Ссылки на проект
+
+IP 51.250.15.37
+Frontend https://front.noctkpuntym.nomoreparties.co
+Backend https://back.noctkpuntym.nomoreparties.co
