+++ app.js
@@ -1,10 +1,9 @@
 const express = require('express');
 const mongoose = require('mongoose');
 const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
 const { errors } = require('celebrate');
-const NotFoundError = require('./errors/not-found-err');
-const { signinValidation, signupValidation } = require('./middlewares/celebrate');
+const limiter = require('./utils/limiter');
+const routers = require('./routes');
 const cors = require('./middlewares/cors');
 
 const { PORT = 3000, DB_URL = 'mongodb://127.0.0.1:27017/bitfilmsdb' } = process.env;
@@ -14,12 +13,7 @@ const app = express();
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 
 app.use(helmet());
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100,
-  standardHeaders: true,
-  legacyHeaders: false,
-});
+
 app.use(limiter);
 
 app.use(cors);
@@ -30,23 +24,12 @@ app.get('/crash-test', () => {
   }, 0);
 });
 
-const usersRouter = require('./routes/users');
-const moviesRouter = require('./routes/movies');
-
-const { createUser, login } = require('./controllers/users');
-const auth = require('./middlewares/auth');
 const errorsMid = require('./middlewares/errors');
 
 app.use(express.json());
-app.use(requestLogger); // подключаем логгер запросов
+app.use(requestLogger);
 
-app.post('/signin', signinValidation, login);
-app.post('/signup', signupValidation, createUser);
-app.use('/users', auth, usersRouter);
-app.use('/movies', auth, moviesRouter);
-app.use((req, res, next) => {
-  next(new NotFoundError({ message: 'Страница не найдена' }));
-});
+app.use(routers);
 
 app.use(errorLogger);
 app.use(errors());
